Simulate trajectory with relation to time (instead of distance)
Trajectory is now a function of t (time) instead of x (distance). This results in the same shape, but has the advantage of being more easily transferable to a ball in motion. The points in `List<Vector2> points` are now spaced according to time, so a ball will be able to animate linearly over these points to follow the trajectory at the correct speed.